1.0.1-PRE-RELEASE Rules handling update!
Fixed the ChatListener error when trying to run rules as the target

Added the strings to the config.yml for the future url commands.

Changed ReviewRulesCommand and JoinLeaveListener to both use "p.preformcommand("rules") instead of having them send a custom msg for it.

Moved the effectsGive() and effectsTake() into RulesFile

Removed RulesListener.java